requirements.txt
logging==0.4.9.6
paho-mqtt==1.5.1
pkg-resources==0.0.0
PyYAML==5.4.1
